Added EventsController, EventsService, separated logic working with EventOccurrence and ScheduledEvents between EventsController and SchedeleController
