+++ CharlieBackend.Business/Services/Interfaces/IScheduleService.cs
@@ -16,18 +16,13 @@ namespace CharlieBackend.Business.Services.Interfaces
 
         public Task<Result<IList<EventOccurrenceDTO>>> GetEventOccurrencesAsync();
 
-        public Task<Result<ScheduledEventDTO>> UpdateScheduledEventByID(long scheduledEvendID, UpdateScheduledEventDto scheduleModel);
-
         public Task<Result<IList<ScheduledEventDTO>>> UpdateEventsRange(ScheduledEventFilterRequestDTO filter, UpdateScheduledEventDto request);
 
         public Task<Result<IList<ScheduledEventDTO>>> GetEventsFiltered(ScheduledEventFilterRequestDTO request);
 
         public Task<Result<EventOccurrenceDTO>> DeleteScheduleByIdAsync(long studentGroupId, DateTime? startDate, DateTime? finishDate);
 
-        public Task<Result<bool>> DeleteConcreteScheduleByIdAsync(long id);
-
         public Task<Result<EventOccurrenceDTO>> UpdateEventOccurrenceById(long eventOccurrenceId, CreateScheduleDto request);
 
-        public Task<Result<ScheduledEventDTO>> GetConcreteScheduleByIdAsync(long eventId);
     }
 }
